fix: additional lint issues after uint64 migration
- Removed unused nolint directives in queue/sharedmain (conversions are now safe)
- Changed if-else to tagged switch for req.eventType (staticcheck QF1003)
- Removed unnecessary uint64() conversion (already uint64)
- Added nolint:unparam for error return (kept for future extensibility)

All tests pass.

+++ pkg/activator/net/throttler_test.go
@@ -254,7 +254,7 @@ func TestThrottlerUpdateCapacity(t *testing.T) {
 				close(rt.done) // Clean up the worker goroutine
 			}()
 
-			rt.containerConcurrency.Store(uint64(tt.containerConcurrency))
+			rt.containerConcurrency.Store(tt.containerConcurrency)
 			rt.numActivators.Store(uint64(tt.numActivators))
 			// Convert 0-based test index to 1-based: 0->1, 1->2, 2->3, etc.
 			// If test specifies 0 (unset), use 1 as default active index
